re-arrange folders and add schema validation
full schema validation for OCDS is not practical b/c (a) it would
make writing unit tests really painful, and (b) it's not the job of
these indicator checks to validate OCDS data. however, to ensure
that the indicator checks work as intended, it is important to
validate inputs against the schema. this approach strikes a balance
between those concerns by validating against the schema but
eliminating required fields