add JSON self-documentation for indicators
with this commit JSDoc is gone (good riddance!) and instead every
indicator self-documents based on the required fields and
preconditions passed into the indicator factory function.

right now the self-documentation only spits out JSON, but it will
be pretty trivial to hook that into a static site script and push
it to gh-pages with a commit hook.