add a factory for collection indicators and implement i004
collection indicators are indicators that need to look at every
record in a collection, e.g. to see supplier collusion across bids,
etc. these are a minority of the indicators that we're creating,
but are important enough that they need proper handling

the current approach sacrifices performance for testability. in a
final implementation, we shoudl probably pass a mongo cursor to
the function instead of buffering the entire collection of releases
into memory only to filter most of them out. however, writing that
in a testable way would requiring mocking the mongo cursor or having
some untested logic, so for the moment focusing on correctness rather
than performance